FO-2511: Splitter opp Typer
