+++ src/view/aktivitet/AktivitetskortPreview.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { Aktivitet, AktivitetTypes, ArenaAktivitet, ArenaAktivitetTypes, DialogData } from '../../utils/Typer';
+import { DialogData } from '../../utils/Typer';
+import { Aktivitet, AktivitetTypes, ArenaAktivitet, ArenaAktivitetTypes } from '../../utils/AktivitetTypes';
 import { Undertekst, Undertittel } from 'nav-frontend-typografi';
 
 import UseFetch from '../../utils/UseFetch';
