+++ src/view/aktivitet/MoteAktivitet.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { formaterDate, getKlokkeslett, getVarighet } from '../../utils/Date';
-import { Aktivitet, KanalTypes } from '../../utils/Typer';
+import { Aktivitet, KanalTypes } from '../../utils/AktivitetTypes';
 import InformasjonElement from './InformasjonElement';
 
 interface PropTypes {
